pyproject.toml
[project]
name = "pp-email-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.1",
    "sendgrid>=6.12.5",
    "sqlalchemy>=2.0.47",
]
